Switch fetching to use event publishing.
